Speedup 'fmdt-detect' by avoiding useless copies.

+++ include/common/fmdt/features.h
@@ -45,10 +45,10 @@ void features_clear_index_ROI_array(ROI_t* ROI_array, const size_t r);
 void features_copy_elmt_ROI_array(const ROI_t* ROI_array_src, ROI_t* ROI_array_dest, const int i_src, const int i_dest);
 void features_copy_ROI_array(const ROI_t* ROI_array_src, ROI_t* ROI_array_dest);
 void features_init_ROI(ROI_t* stats, int n);
-void features_extract(const uint32_t** img, const int i0, const int i1, const int j0, const int j1,
-                      const size_t n_ROI, ROI_t* ROI_array);
+void features_extract(const uint32_t** img, const int i0, const int i1, const int j0, const int j1, const size_t n_ROI,
+                      ROI_t* ROI_array);
 // void features_filter_surface(ROI_t* ROI_array, uint32_t** img, uint32_t threshold_min, uint32_t threshold_max);
-void features_merge_HI_CCL_v2(const uint32_t** M, const uint32_t** HI_in, uint32_t** HI_out, const int i0, const int i1,
+void features_merge_HI_CCL_v2(const uint32_t** M, const uint8_t** HI_in, uint8_t** HI_out, const int i0, const int i1,
                               const int j0, const int j1, ROI_t* ROI_array, const uint32_t S_min, const uint32_t S_max);
 void features_shrink_ROI_array(const ROI_t* ROI_array_src, ROI_t* ROI_array_dest);
 double features_compute_mean_error(const ROI_t* stats);
